renamed exit beta to modal
